Data taken from data.sandiego.gov
Traffic collisions - people and vehicles involved

"Traffic collision reports recorded by the San Diego Police Department with additional details on people and vehicles involved.
Generally a report is not taken for property damage-only collisions that do not involve hit & run or DUI. The California Highway Patrol is responsible for handling collisions occurring on the freeway.

A single collision may involve multiple people and/or vehicles. This dataset contains one row per person involved in a collision. Each collision is uniquely identified in the report_id field.

Throughout 2018 and 2019, a new collision reporting process was introduced that resulted in more detailed data collection. Data fields listed in the Data Dictionary that include ** in the description may have incomplete data values for history records."

https://data.sandiego.gov/datasets/police-collisions-details/